requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appscript==1.3.0
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.2
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonformer==0.12.0
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.74.1
packaging==24.2
pfzy==0.3.4
pillow==11.2.1
prompt_toolkit==3.0.51
pydantic==2.11.3
pydantic_core==2.33.1
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==78.1.0
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
